feat(ios): add close
Update iOS framework to the latest version.
Implement close feature. Instead of managing the opened router, just manage the opened view controller and, when 'close' is triggered, dismiss it. The view itself will take care of everything related to the operation.
Add a close error, when there's no view to close.
Add a 'handleResult' to manage the External Browser and Web View event reaction. To make every operation synchronised, the events should only be triggered when the display view is completed. That's why we add the `present` method.

References: https://outsystemsrd.atlassian.net/browse/RMET-3427